Dispatch event to allow apps to load their script for collections
