+++ apps/files_sharing/webpack.js
@@ -5,6 +5,7 @@ module.exports = {
 	entry: {
 		'additionalScripts': path.join(__dirname, 'src', 'additionalScripts.js'),
 		'files_sharing': path.join(__dirname, 'src', 'files_sharing.js'),
+		'collaboration': path.join(__dirname, 'src', 'collaborationresourceshandler.js'),
 	},
 	output: {
 		path: path.resolve(__dirname, './js/dist/'),
